webapp/backend/requirements.txt
uvicorn
fastapi
pandas
torch
uuid
pybase64
mlxtend
spacy==3.7.5
python-dotenv
ollama
transformers
rq
markdown_crawler
markdownify
langdetect
deep-translator
rank_bm25
py2neo
pymupdf
gensim
nltk
sentencepiece
python-multipart
sentence_transformers
neo4j
fastopic
pydantic
pydantic-settings